FIX: dependencies cycle fix
