+++ backend/app/models.py
@@ -1,7 +1,7 @@
 from flask_login import UserMixin
 from werkzeug.security import check_password_hash, generate_password_hash
 
-from . import db
+from backend.extensions import db
 
 
 class Position(db.Model):
