+++ backend/app/routes.py
@@ -1,7 +1,7 @@
 from flask import Blueprint, jsonify, request
 from flask_login import current_user, login_required
 
-from . import db
+from backend.extensions import db
 from .models import Position, User
 
 bp = Blueprint("main", __name__)
